requirements.txt
python-telegram-bot[job-queue]==21.4
psycopg[binary]==3.2.9
python-dotenv==1.0.1
